Initial changes.
This part is only has visual and some funcionality, so far i've only accomplished the comunication between the php files that manage the login and register. But it does nothing so far, it only receives the information via AJAX; but, it looks very good for me. I've never made a desing look so good so far, so this really makes me happy and Bootstrap 4 it's a great tool for layout mostly.